Move the sudo required tests into a single test
