+++ e2e/schemas/createSchema.test.ts
@@ -6,9 +6,7 @@ import { AVRO_GRAPH_CHANGE } from "./fixtures/avroGraphChangeSchemaType";
 import { KeyringPair } from "@polkadot/keyring/types";
 import { ExtrinsicHelper } from "../scaffolding/extrinsicHelpers";
 import { createKeys, createAndFundKeypair, assertExtrinsicSuccess } from "../scaffolding/helpers";
-import { u16 } from "@polkadot/types";
-import { getFundingSource, getSudo } from "../scaffolding/funding";
-import { isTestnet } from "../scaffolding/env";
+import { getFundingSource } from "../scaffolding/funding";
 
 describe("#createSchema", function () {
   let keys: KeyringPair;
@@ -69,32 +67,4 @@ describe("#createSchema", function () {
     assert.notEqual(createSchemaEvent, undefined);
   });
 
-  it("should fail to create non itemized schema with AppendOnly settings", async function () {
-    if (isTestnet()) this.skip();
-
-    const sudoKey = getSudo().keys;
-    const ex = ExtrinsicHelper.createSchemaWithSettingsGov(keys, sudoKey, AVRO_GRAPH_CHANGE, "AvroBinary", "Paginated", "AppendOnly");
-    await assert.rejects(ex.sudoSignAndSend(), {
-      name: 'InvalidSetting'
-    });
-  });
-
-  it("should not fail to create itemized schema with AppendOnly settings", async function () {
-    if (isTestnet()) this.skip();
-
-    const sudoKey = getSudo().keys;
-    const createSchema = ExtrinsicHelper.createSchemaWithSettingsGov(keys, sudoKey, AVRO_GRAPH_CHANGE, "AvroBinary", "Itemized", "AppendOnly");
-    const [event] = await createSchema.sudoSignAndSend();
-    assert.notEqual(event, undefined);
-    let itemizedSchemaId: u16 = new u16(ExtrinsicHelper.api.registry, 0);
-    if (event && createSchema.api.events.schemas.SchemaCreated.is(event)) {
-      itemizedSchemaId = event.data.schemaId;
-    }
-    assert.notEqual(itemizedSchemaId.toNumber(), 0);
-    let schema_response = await ExtrinsicHelper.getSchema(itemizedSchemaId);
-    assert(schema_response.isSome);
-    let schema_response_value = schema_response.unwrap();
-    let schema_settings = schema_response_value.settings;
-    assert.notEqual(schema_settings.length, 0);
-  });
 })
